+++ lib/ace/http2/error.ex
@@ -1,6 +1,5 @@
 defmodule Ace.HTTP2.Errors do
-  @moduledoc """
-  """
+  @moduledoc false
 
   @defined [
     {0x0, :no_error},